requirements.txt
# CONFIGURATIOn
pyyaml
# REDDIT API
praw
